requirements.txt
# Core
fastapi>=0.95.1
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0

# Database & vectors
psycopg2-binary>=2.9.5
pgvector>=0.4.0

# OpenAI client
openai>=1.0.0

# HTTP client for tests
httpx>=0.24.0

# Testing
pytest>=7.4.0
